Move elevator up and down
This adds the functionality to move the elevator up and down.  It adds
a timer delay to simulate actual time taking it to move.  In the future
this will be used by some asyncronous code.  It should also be changing
self.currrent_floor each time it moves but that will come later.