Limit elevator floor movement to max and min
This limits the floor movements to the max and min floors.  If the move
function is called with a floor above the maximum floor count it will
raise a MaxFloorError.  If the move function is called with a floor
index of less then the min floor count then a MinFloorError will be
raised.